# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Collaborative Kanban board application with real-time updates",
        "title": "CollabBoard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/boards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all boards where the authenticated user is a member or owner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "List user's boards",
                "responses": {
                    "200": {
                        "description": "List of boards",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/boards.BoardResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new Kanban board with the authenticated user as owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Create a new board",
                "parameters": [
                    {
                        "description": "Board creation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/boards.CreateBoardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Board created successfully",
                        "schema": {
                            "$ref": "#/definitions/boards.BoardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/boards/by-role/{role}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get boards where the user has a specific role (owner or member)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "List boards by role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role filter (owner or member)",
                        "name": "role",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of boards",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/boards.BoardResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid role parameter",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/boards/{boardId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a specific board",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Get board by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Board information",
                        "schema": {
                            "$ref": "#/definitions/boards.BoardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not a board member",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update board information (only board owners can update)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Update board",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Board update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/boards.UpdateBoardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Board updated successfully",
                        "schema": {
                            "$ref": "#/definitions/boards.BoardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - only owners can update",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a board (only board owners can delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boards"
                ],
                "summary": "Delete board",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Board deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/boards.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - only owners can delete",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/boards/{boardId}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all members of a specific board",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board Members"
                ],
                "summary": "List board members",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of board members",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/boards.BoardMemberResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not a board member",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a user as a member to a board (only board owners can add members)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board Members"
                ],
                "summary": "Add board member",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Member details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/boards.AddMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Member added successfully",
                        "schema": {
                            "$ref": "#/definitions/boards.BoardMemberResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - only owners can add members",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/boards/{boardId}/members/leave": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allow a user to leave a board they are a member of",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board Members"
                ],
                "summary": "Leave board",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully left the board",
                        "schema": {
                            "$ref": "#/definitions/boards.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - cannot leave board",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/boards.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "boards.AddMemberRequest": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "role": {
                    "type": "string",
                    "example": "member"
                },
                "userId": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "boards.BoardResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "My Project Board"
                },
                "ownerId": {
                    "type": "integer",
                    "example": 1
                },
                "role": {
                    "type": "string",
                    "example": "owner"
                }
            }
        },
        "boards.CreateBoardRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "My Project Board"
                }
            }
        },
        "boards.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Board not found"
                }
            }
        },
...
        "boards.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Board deleted"
                }
            }
        },
        "boards.UpdateBoardRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Updated Board Name"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}